refactor : Cars - postion 관련 로직 매핑
